Moved some stuff Moved abilities to core/abilities Fixed visibility of BlockWatcher
