+++ src/main/java/net/laboulangerie/laboulangeriemmo/api/player/MmoPlayerListener.java
@@ -9,7 +9,7 @@ import org.bukkit.event.player.PlayerJoinEvent;
 import org.bukkit.event.player.PlayerQuitEvent;
 
 import net.laboulangerie.laboulangeriemmo.LaBoulangerieMmo;
-import net.laboulangerie.laboulangeriemmo.core.MarkedBlocksManager;
+import net.laboulangerie.laboulangeriemmo.core.abilities.mining.MarkedBlocksManager;
 import net.laboulangerie.laboulangeriemmo.events.MmoPlayerBreakBlockEvent;
 import net.laboulangerie.laboulangeriemmo.events.PlayerEarnsXpEvent;
