+++ src/main/java/net/laboulangerie/laboulangeriemmo/core/abilities/thehunter/FireBow.java
@@ -1,4 +1,4 @@
-package net.laboulangerie.laboulangeriemmo.abilities.thehunter;
+package net.laboulangerie.laboulangeriemmo.core.abilities.thehunter;
 
 import java.util.Random;
