+++ src/main/java/net/laboulangerie/laboulangeriemmo/core/abilities/woodcutting/Strip.java
@@ -1,20 +1,23 @@
-package net.laboulangerie.laboulangeriemmo.abilities.woodcutting;
+package net.laboulangerie.laboulangeriemmo.core.abilities.woodcutting;
 
 import org.bukkit.Location;
 import org.bukkit.Material;
 import org.bukkit.Tag;
 import org.bukkit.block.Block;
+import org.bukkit.entity.Player;
 import org.bukkit.event.Event;
-import org.bukkit.event.block.BlockBreakEvent;
 import org.bukkit.scheduler.BukkitRunnable;
 
 import net.laboulangerie.laboulangeriemmo.LaBoulangerieMmo;
 import net.laboulangerie.laboulangeriemmo.api.ability.AbilityArchetype;
 import net.laboulangerie.laboulangeriemmo.api.ability.AbilityExecutor;
 import net.laboulangerie.laboulangeriemmo.api.ability.AbilityTrigger;
+import net.laboulangerie.laboulangeriemmo.core.combo.ComboKey;
+import net.laboulangerie.laboulangeriemmo.core.combo.KeyStreak;
+import net.laboulangerie.laboulangeriemmo.events.ComboCompletedEvent;
 
-public class Timber extends AbilityExecutor {
-    public Timber(AbilityArchetype archetype) {
+public class Strip extends AbilityExecutor {
+    public Strip(AbilityArchetype archetype) {
         super(archetype);
     }
 
@@ -54,32 +57,32 @@ public class Timber extends AbilityExecutor {
 
     @Override
     public AbilityTrigger getAbilityTrigger() {
-        return AbilityTrigger.BREAK;
+        return AbilityTrigger.COMBO;
     }
 
     @Override
     public boolean shouldTrigger(Event baseEvent) {
-        BlockBreakEvent event = (BlockBreakEvent) baseEvent;
-        Block block = event.getBlock();
+        ComboCompletedEvent event = (ComboCompletedEvent) baseEvent;
+        Block block = event.getPlayer().getTargetBlock(5);
 
-        return block != null && Tag.LOGS.isTagged(block.getType()) && !(block.hasMetadata("laboulangerie:placed"));
+        return new KeyStreak(ComboKey.LEFT, ComboKey.LEFT, ComboKey.LEFT).match(event.getKeyStreak()) && block != null && Tag.LOGS.isTagged(block.getType()) && !(block.hasMetadata("laboulangerie:placed"));
     }
 
     @Override
     public void trigger(Event baseEvent, int level) {
-        BlockBreakEvent event = (BlockBreakEvent) baseEvent;
-        Block initBlock = event.getBlock();
+        ComboCompletedEvent event = (ComboCompletedEvent) baseEvent;
+        Block initBlock = event.getPlayer().getTargetBlock(5);
         initType = initBlock.getType();
         initLocation = initBlock.getLocation();
-        breakNeighbours(initBlock);
+        stripNeighbours(initBlock, event.getPlayer());
     }
 
-    private void breakNeighbours(Block block) {
+    private void stripNeighbours(Block block, Player player) {
         new BukkitRunnable() {
             @Override
             public void run() {
                 Location loc = block.getLocation();
-                for (int[] coordinate : Timber.relCoordinates) {
+                for (int[] coordinate : Strip.relCoordinates) {
                     Location neighbourLoc = loc.clone().add(coordinate[0], coordinate[1], coordinate[2]);
                     Block neighbour = neighbourLoc.getBlock();
 
@@ -87,10 +90,9 @@ public class Timber extends AbilityExecutor {
                             && neighbour.getY() >= initLocation.getBlockY()
                             && Math.abs(neighbour.getX() - initLocation.getBlockX()) <= range
                             && Math.abs(neighbour.getZ() - initLocation.getBlockZ()) <= range) {
-                        // Drop the item and spawn block particles
-                        neighbour.breakNaturally(null, true);
+                        block.setType(Material.getMaterial("STRIPPED_" + block.getType().toString()));
                         // Break neighbours of neighbour recursively
-                        breakNeighbours(neighbour);
+                        stripNeighbours(neighbour , player);
                     }
                 }
             }
